third_party/WebKit/LayoutTests/fast/html/imports/import-defaultView-expected.txt

PASS importLink.import.defaultView is null
PASS xhr.response.defaultView is null
PASS createdDoc.defaultView is null
PASS successfullyParsed is true

TEST COMPLETE
